feat: API 관련 상수 config 파일로 분리
